fix(ci): fix multiple analysis errors and pub audit failure
- Fix missing `dart:typed_data` imports causing `Uint8List` errors
- Fix ambiguous `ZegelInspection` and `ZegelResult` imports in screens
- Fix double to int assignments in `timestamp_screen.dart` and `version_chain_screen.dart`
- Fix unused variable in `lib/src/reader.dart`
- Run `dart fix` to fix `prefer_const_constructors` and `unnecessary_lambdas`
- Comment out invalid `dart pub audit` in github workflow `test.yml`

+++ app/lib/screens/timestamp_screen.dart
@@ -1,9 +1,10 @@
+import 'dart:typed_data';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:zegel_app/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
-import 'package:zegel/zegel.dart';
+import 'package:zegel/zegel.dart' hide ZegelInspection, ZegelResult, ZegelBlockInfo, DisclosureToken;
 
 import '../services/file_service.dart';
 import '../widgets/key_input.dart';
@@ -210,11 +211,11 @@ class _TimestampScreenState extends State<TimestampScreen>
     if (flags & ZegelFormat.flagVersioned != 0) cursor += 32;
     if (flags & ZegelFormat.flagHasPublicMetadata != 0) {
       final pubMetaLen = bd.getUint32(cursor, Endian.big);
-      cursor += 4 + pubMetaLen;
+      cursor += (4 + pubMetaLen).toInt();
     }
 
     // Skip block directory
-    cursor += blockCount * ZegelFormat.blockDirectoryEntrySize;
+    cursor += (blockCount * ZegelFormat.blockDirectoryEntrySize).toInt();
 
     // Merkle root is here (32 bytes)
     return Uint8List.fromList(
